Edit and Delete Account APIs
Main addition is the above APIs. They take in a "userId" to retrieve the user. The Edit API can be used to edit the account's First and Last Name, Email, and Password (must confirm old password). Delete deletes the account; there is also code that has been commented out for extra security.
Other changes are that the auto-format feature was used in Visual Studio Code.