Several Notification-related Endpoints
4 endpoints have been added: Create, Get, Delete 1, Delete All Notifications. They operate under this format
```http://localhost:4000/api/users/:UserId/notifications```
where it finds a user in Users and pulls their notifications. The one for Delete 1 also needs to have this added to the end
```/:notificationIndex```
where ":notificationIndex" is the position of whichever notification you're looking for in the array (so array rules apply, where 0 is the first one).