Hamburger Menu Unit Tests (API)
Added 53 tests for the APIs of the Hamburger Menu. These can be found inside the hamburgapi.test.js file.
Also updated package.json to add a piece of code that lets us on only test the added tests. How this works is running "npm test:x" where x is whatever we are testing. My prior unit tests for login and register can be run with x being "auth" and these ones for the Hamburger Menu can be run with "hamburg". Of course, running "npm run test" by itself will also just run all of the tests.