Massive Notifications Overhaul + Meetup Addition
Due to the Meetup Notifications creation process, a massive overhaul was needed for Notifications. As such, a Notifications Schema has been created. APIs have been adjusted and created to compensate for this. As well, the Unit Tests have been updated to deal with it. For more information on all of these, look at the comments included in the code itself.
As for the Meetup process itself, the system we settled on has been baked into the "Create Notification Endpoint". We will have the creator of the notification send it to the receiver, who will then have several options *on the front end* for how to respond. For the most part, the return notification will usually just be a normal notification. However, if it's asking for a change in the meet, it will send as a Meetup Notification as well.