Updated Mobile Account Settings Page (+ Code Readability)
Mobile App's Account Settings has seen a pretty neat overhaul. It also has some things written as TODO in the comments (such as adding API and actually making Dark Mode and such work).
I also used the auto-adjust feature in Visual Studio Code to hopefully help with the readability of the code in all the .dart files.
As usual, feel free to revert the changes if anything is wrong.