Mobile App Matches Figma
Made the mobile app more closely appear like the Figma design. This involved a number of changes: mainly with the creation of AppColors, a class that has all of our color palettes, inside of main.dart. However, a lot of changes were also made. If any of these changes cause conflicts with development however, roll it back and disregard; we'll find another way to handle things.