[Mobile] Your Reports Section (Pass 1)
Adding in the Your Reports section into the app. This is done by creating a new file "your_reports.dart" and linking it to home.dart. Currently, it looks nice but I haven't figured out how to get the filtering to work correctly (assuming we need filtering for this). It might be with me using the wrong API but I can't find a way to filter by status, category, and name. Anyways, Edit also doesn't work yet but Delete does! Will get to work on this more tomorrow.