[Mobile] Your Reports Section (Pass 2)
Your Reports is now done! From the last time, the filters for the search actually works (I remembered that we had asked for a specific change to one of the APIs so I swapped to using that one instead). As well, the Edit function is confirmed to work (changes on the mobile version reflected on web). It was made by creating a new .dart file. Using the Edit function highlighted something I believe should be addressed however (it regards the map). More discussion about this will be had over the chats.
As an extra note, I hit Save on notifications.dart and it auto-formatted it, so don't worry about that.