Inbox/Notifications Section [Mobile]
Updated the Inbox/Notifications Section of the mobile end to function. Displays notifications or doesn't if you don't have any. The Top has options to Mark All Read or Delete All. The Notif itself shows if its been read or not, and if not has a Mark Read option. Delete is always present to delete the notification. When opening the Notification, it displays the details as well as the related item's details if applicable. If it's a MeetUp Notification, it also displays the Meeting Location and Time as well the three different Meetup Responses.