Meetup return to Sender [Notification on Mobile]
Upon the user selecting one of the 3 options for Meetup, 3 things will happen: the Meetup notification will be deleted, the user who made the decision will get a notification back to them explaining what their decision was, and the user who initially sent the Meetup notification will be informed of what the decision was.
When Accepting or Contesting, the location and date will be contained within the text of the notification sent back.