IT'S ABOUT TIME (ft. Notif Consistency)
AFTER AGONZING HOURS, I have finally managed to get the time seen everywhere be consistent. In addition to this, I have updated the text of the notifications in both the web and mobile ends to be consistent with each other. The last thing is that I checked the flags for isMeetup: everything should be false outside of the initial Found notification and the Contests sent as a response on both the mobile and web ends.